+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0-alpha.4] - 2024-01-24
+
+### Fixed
+
+- Logic to assess if we are on a pre-release
+
+### Added
+
+- Allow passing a timestamp option to `generate` and `get-version`
+
 ## [0.3.0-alpha.3] - 2024-01-23
 
 ### Added